SUM-60 & SUM-62
volume slider is now persistent between games and scenes, UI elements are appropriately hidden or shown when necessary.  Start of a script which handles the end of the game, but that still needs some work.

also using TMPro does not work for me, maybe it works for other people idk but without that idk how to make the UI text that shows the accurate number for the volume slider